Fix and remove GLAD bin dir files

+++ Infinite/src/Infinite/Renderer/Buffer.h
@@ -108,7 +108,7 @@ namespace Infinite {
 		virtual void Bind() const = 0;
 		virtual void Unbind() const = 0;
 
-		virtual void SetData(const void* data, uint32_t size) = 0;
+		virtual void SetData(const void* data, size_t size) = 0;
 
 		virtual void SetLayout(const BufferLayout& layout) = 0;
 		virtual const BufferLayout& GetLayout() const = 0;
